refactor: move test files in /src
