+++ src/Box/index.spec.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import {render, cleanup} from '@testing-library/react';
-import Wrapper from '../helper/wrapper';
-import {ThemeSectionError} from '../../src/ThemeSection';
-import Box from '../../src/Box';
+// import Wrapper from '../../test/helper/wrapper';
+import {ThemeSectionError} from '..//ThemeSection';
+import Box from '.';
 
 describe('Box', () => {
 	afterEach(cleanup);
