+++ src/Flex/index.spec.js
@@ -1,15 +1,15 @@
 import React from 'react';
 import {render, cleanup} from '@testing-library/react';
-import Wrapper from '../helper/wrapper';
-import {ThemeSectionError} from '../../src/ThemeSection';
-import Switch from '../../src/Switch';
+// import Wrapper from '../../test/helper/wrapper';
+import {ThemeSectionError} from '..//ThemeSection';
+import Flex from '.';
 
-describe('Switch', () => {
+describe('Flex', () => {
 	afterEach(cleanup);
 
 	it('throws ThemeSectionError if rendered outside a <ThemeSection />', () => {
 		expect(() => {
-			render(<Switch />).toThrow(ThemeSectionError);
+			render(<Flex />).toThrow(ThemeSectionError);
 		});
 	});
 });
