+++ src/PopOver/index.spec.js
@@ -1,15 +1,15 @@
 import React from 'react';
 import {render, cleanup} from '@testing-library/react';
-import Wrapper from '../helper/wrapper';
-import {ThemeSectionError} from '../../src/ThemeSection';
-import Text from '../../src/Text';
+// import Wrapper from '../../test/helper/wrapper';
+import {ThemeSectionError} from '../ThemeSection';
+import PopOver from '.';
 
-describe('Text', () => {
+describe('PopOver', () => {
 	afterEach(cleanup);
 
 	it('throws ThemeSectionError if rendered outside a <ThemeSection />', () => {
 		expect(() => {
-			render(<Text />).toThrow(ThemeSectionError);
+			render(<PopOver />).toThrow(ThemeSectionError);
 		});
 	});
 });
