--- test/PopOver/index.spec.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import {render, cleanup} from '@testing-library/react';
-import Wrapper from '../helper/wrapper';
-import {ThemeSectionError} from '../../src/ThemeSection';
-import PopOver from '../../src/PopOver';
-
-describe('PopOver', () => {
-	afterEach(cleanup);
-
-	it('throws ThemeSectionError if rendered outside a <ThemeSection />', () => {
-		expect(() => {
-			render(<PopOver />).toThrow(ThemeSectionError);
-		});
-	});
-});
